+++ CHANGELOG.md
@@ -1,0 +1,4 @@
+[2.0.0]
+
+- Renamed gem to motion-expect (based on discussion with original owner - https://github.com/ijpiantanida/bacon-expect/issues/4)
+- Updated for RubyMotion 6
