+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in bacon-expect-wizard.gemspec
+# Specify your gem's dependencies in motion-expect.gemspec
 gemspec
