+++ README.md
@@ -1,7 +1,4 @@
-**UNMAINTAINED PROJECT**   
-**I stopped developing in RubyMotion long time ago and can no longer maintain the project.**
-
-# BaconExpect [![Build Status](https://travis-ci.org/ijpiantanida/bacon-expect.svg?branch=master)](https://travis-ci.org/ijpiantanida/bacon-expect) [![Code Climate](https://codeclimate.com/github/ijpiantanida/bacon-expect.png)](https://codeclimate.com/github/ijpiantanida/bacon-expect)
+# MotionExpect [![Build Status](https://travis-ci.org/rubymotion-community/motion-expect.svg?branch=master)](https://travis-ci.org/rubymotion-community/motion-expect) [![Code Climate](https://codeclimate.com/github/rubymotion-community/motion-expect.png)](https://codeclimate.com/github/rubymotion-community/motion-expect)
 
 Bring [RSpec 3.0 expect syntax](https://www.relishapp.com/rspec/rspec-expectations/v/3-0/docs/built-in-matchers) to RubyMotion's MacBacon.
 ```ruby
@@ -12,10 +9,10 @@ Expect syntax allows a cleaner implementation because it doesn't need to monkey-
 
  # Installation
 ```
-gem install bacon-expect
+gem install motion-expect
 
 # or in Gemfile
-gem 'bacon-expect'
+gem 'motion-expect'
 ```
 
 # Usage
