--- lib/bacon-expect/bacon_context.rb
@@ -1,4 +1,0 @@
-class Bacon::Context
-  include BaconExpect::Expectation::BaconContext
-  include BaconExpect::Matcher::BaconContext
-end
