--- lib/bacon-expect/matchers/eql.rb
@@ -1,7 +1,0 @@
-module BaconExpect; module Matcher
-  class Eql < SingleMethod
-    def initialize(value)
-      super(:eql?, value)
-    end
-  end
-end; end
