+++ lib/motion-expect.rb
@@ -3,7 +3,7 @@ unless defined?(Motion::Project::Config)
 end
 
 Motion::Project::App.setup do |app|
-  Dir.glob(File.join(File.dirname(__FILE__), 'bacon-expect/**/*.rb')).reverse.each do |file|
+  Dir.glob(File.join(File.dirname(__FILE__), 'motion-expect/**/*.rb')).sort.reverse.each do |file|
     app.spec_files << file
   end
 end
