+++ lib/motion-expect/bacon_context.rb
@@ -1,0 +1,4 @@
+class Bacon::Context
+  include MotionExpect::Expectation::BaconContext
+  include MotionExpect::Matcher::BaconContext
+end
