+++ lib/motion-expect/exceptions.rb
@@ -1,4 +1,4 @@
-module BaconExpect
+module MotionExpect
   class FailedExpectation < RuntimeError; end
   class InvalidMatcher < RuntimeError; end
 end
