+++ lib/motion-expect/matcher/be.rb
@@ -1,4 +1,4 @@
-module BaconExpect; module Matcher
+module MotionExpect; module Matcher
   class Be < SingleMethod
     def initialize(value)
       super(:equal?, value)
