+++ lib/motion-expect/matcher/be_false.rb
@@ -1,4 +1,4 @@
-module BaconExpect; module Matcher
+module MotionExpect; module Matcher
   class BeFalse
     def matches?(value)
       !value
