+++ lib/motion-expect/matcher/be_generic.rb
@@ -1,4 +1,4 @@
-module BaconExpect; module Matcher
+module MotionExpect; module Matcher
   class BeGeneric < SingleMethod
     def initialize(method_name, *values)
       super("#{method_name}?", *values)
