+++ lib/motion-expect/matcher/be_true.rb
@@ -1,4 +1,4 @@
-module BaconExpect; module Matcher
+module MotionExpect; module Matcher
   class BeTrue
     def matches?(value)
       value == true
