+++ lib/motion-expect/matcher/eql.rb
@@ -1,0 +1,9 @@
+module MotionExpect
+  module Matcher
+    class Eql < SingleMethod
+      def initialize(value)
+        super(:eql?, value)
+      end
+    end
+  end
+end
