+++ lib/motion-expect/matcher/include.rb
@@ -1,4 +1,4 @@
-module BaconExpect; module Matcher
+module MotionExpect; module Matcher
   class Include
     def initialize(*values)
       @values = *values
