+++ lib/motion-expect/matcher/respond_to.rb
@@ -1,4 +1,4 @@
-module BaconExpect; module Matcher
+module MotionExpect; module Matcher
   class RespondTo
     def initialize(method_name)
       @method_name = method_name
