+++ spec_app/Gemfile
@@ -2,4 +2,4 @@ source 'https://rubygems.org'
 
 gem 'rake'
 # Add your dependencies here:
-gem 'bacon-expect', :path => ".."
+gem 'motion-expect', :path => ".."
