+++ spec_app/spec/helpers/bacon_context.rb
@@ -1,5 +1,5 @@
 class Bacon::Context
   def expect_failure(fail_message = "", &block)
-    expect(&block).to raise_error(BaconExpect::FailedExpectation, fail_message)
+    expect(&block).to raise_error(MotionExpect::FailedExpectation, fail_message)
   end
 end
